requirements.txt
openai>=1.12.0
pydantic>=2.0
python-dotenv>=1.0
playwright>=1.42
mysql-connector-python>=8.0
beautifulsoup4>=4.12 